Update patch set 4
Patch Set 4:

Right, so we see that the "out of bounds" tests failed.
One problem you're going to encounter is that, when the rectangles are out of bounds, they need to be computed properly, not only because scaling may require filtering, but also because the source may not start exactly on a pixel, the zone to copy from might be [0.73f, 25.21f], for example, and then filtering is used to yield the proper result. It's going to be hard to pre-clamp anything. It's probably easier to guard the few places where integer overflow can happen against it, but I'm open to any solution that doesn't break the logic mentioned above.

Patch-set: 4
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>