Update patch set 6
Patch Set 6: Code-Review+1

I just tested it on my macbook and it's interesting, it appears to make the FPS go up, but not all the time. It used to be a fairly constant ~20 FPS, now it's ~20 FPS when both balloons are behind the glass and ~30 FPS when they are offscreen.

Patch-set: 6
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+1