Remove idle Blitter fallbacks
For Vulkan we must support all blittable formats natively, so there's
no fallback to C++ code. Also, if by error an unsupported format is
used we don't have to immediately abort routine generation but can
continue with something that's benign.

This change simplifies the code, makes it more readable/maintainable,
and marks other formats as unsupported instead of unimplemented.

Bug: b/138944025
Change-Id: I4d246cdd280c5a352f18341996c92111a5144d96