Omit null initialization of sampler cache pointers
The last-used sampler routine can be reused when both the image and
sampler descriptors are the same. To prevent a match on uninitialized
cache keys we previously initialized them both to nullptr (as well as
the function pointer itself), at the entry of the shader routine.
This change reduces it to just initializing the image descriptor. This
suffices since all image instructions use an image descriptor
(but not all use a sampler descriptor).

Bug: b/137649247
Change-Id: I62112f12a58ddefb6e56bddab8e3fdfa76c5b756