Update patch set 3
Patch Set 3: Verified+1

(1 comment)

> Patch Set 1: Verified+1 Code-Review+2
> 
> (1 comment)
> 
> As I mentioned in my CL that added this, I'm not fond of floating variables, but okay.
> Note that there is other code pinning variables down just for sanities sake: https://github.com/google/swiftshader/blob/master/src/Pipeline/PixelRoutine.cpp#L40-L45
> 
> In previous jobs, I'd recommend having this stuff conditionally initialized based on a build flag. That way you keep the perf gains, and when stuff breaks nondeterministically, you can flip the initialize-vars-to-some-known-value flag and hopefully have things become reproducible.
> This might be best done at the reactor level though.

Yes, we should do that, as well as adding several other flags for helping locate bugs in either SwiftShader or applications. Made some notes in the 2.0 doc.

FWIW note that initializing the inputs is for matching what most drivers do. Quite a few GL apps relied on it and some devs even thought it was part of the shader linking spec. So rather than demanding all these (internal and external) apps to be fixed we just adopted this de-facto standard. Note that it doesn't negatively impact spec-compliant apps, aside from shader compilation time. Since we're still early in the game for Vulkan, and in a great position to be used for testing by many apps, we may want to leave it uninitialized and instead report the issue...

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Verified=+1