Update patch set 1
Patch Set 1:

It's interesting that this is apparently the only pixel processing state we use for specializing the routine that's not part of the pipeline state. But I'd rather not lose the ability to optimize for that. The majority of draw calls don't need occlusion counting, and I suspect other pipeline stages have more such state (though I could be mistaken since nothing immediately pops to mind).

Anyway, this nicely illustrates that we should probably ensure the legacy 'context' only contains this additional state, and not any pipeline state, so we have a better overview of it and can opt to not optimize for it so we can JIT at pipeline creation time, and only if the code is hot specialize for it.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>