Update patch set 1
Patch Set 1:

I'd rather not support this for now if we can avoid it. We're not even 100% line rasterization conformant yet (Vulkan is more demanding/specific than OpenGL ES).

Also arguably in a low-level API like Vulkan it should be the application's responsibility to implement wireframe rendering. Probably in a geometry shader, although we don't intend to support that any time soon either.

Patch-set: 1
CC: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>