Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'd rather not support this for now if we can avoid it. We're not even 100% line rasterization conformant yet (Vulkan is more demanding/specific than OpenGL ES).

This passes conformance for the tests that I found - were stricter tests added in a later dEQP?

> Also arguably in a low-level API like Vulkan it should be the application's responsibility to implement wireframe rendering. Probably in a geometry shader, although we don't intend to support that any time soon either.

If it is part of the spec, there will be apps that use the feature. It seems like we're going to limit the number of potential clients if we refuse to implement parts of the spec for personal opinions.

Patch-set: 1
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>