[vulkan] Bleeding-edge update Vulkan headers for Fuchsia
This adds a new header <vulkan/vulkan_fuchsia_extras.h> which
provides Fuchsia-specific definitions that are not part of the
upstream Khronos headers yet, and are still only available from
the Fuchsia platform tree.

Putting these into a separate header should make comparisons
and updates to the rest of the Khronos headers much easier in
the future. Once everything has been upstreamed properly, the
extra header will be removed.

NOTE: vulkan.hpp is not updated because doing so is tricky and this
      header is not included by any SwiftShader code.

Change-Id: Ifac0f4fde305156dfe6b349935bab321b2f21622