[vulkan] Bleeding-edge update Vulkan headers for Fuchsia
DO NOT SUBMIT: These are not upstreamed to Khronos yet.

This updates <vulkan/vulkan_core.h> and <vulkan/vulkan_fuchsia.h>
with Fuchsia-specific definitions that are not part of the upstream
Khronos headers yet, and are still only available from the Fuchsia
platform tree.

As such, this CL should not be submitted to the SwiftShader tree,
since final values will likely be slightly different after the
Fuchsia upstreaming process completes.

NOTE: vulkan.hpp is not updated because doing so is tricky and this
      header is not included by any SwiftShader code.