EXPERIMENTAL: [memfd] Create support class for Linux memfd-backed region.
DO NOT SUBMIT: This is a cherry-pick of
https://swiftshader-review.googlesource.com/c/SwiftShader/+/35128
that is only used to allow CLs on top of it to exist in a different
branch (i.e. avoid complicated git merges while reviewing WIP
CLs on Gerrit).

This creates a new helper class to allocate memfd-backed shared
memory regions with the Linux kernel. The feature has been
supported since at least Linux 3.17, which is sufficient for
a large number of desktop Linux and Android devices.

However, only recent version of GLibc or the Android C library
expose memfd_create(), so keep implementing it as a direct syscall.

NOTE: There is a similar function named memfd_create() inside
      of src/Reactor/ExecutableMemory.cpp, but it is easier to
      keep them separate for now to avoid introducing subtle
      dependency issues in the CMakeLists.txt file.

This will used in future CLs that implement Vulkan external
semaphores and externam memory on Linux and Android for SwiftShader.