[memfd] Create support class for Linux memfd-backed region.
This creates a new helper class to allocate memfd-backed shared
memory regions with the Linux kernel. The feature has been
supported since at least Linux 3.17, which is sufficient for
a large number of desktop Linux and Android devices.

However, only recent version of GLibc or the Android C library
expose memfd_create(), so keep implementing it as a direct syscall.

This will used in future CLs that implement external semaphores
and externam memory on Linux and Android for SwiftShader.

See https://swiftshader-review.googlesource.com/c/SwiftShader/+/35129