[vulkan] External Memory and Semaphore support for Linux and Android.
This CL adds implementations of external memory and external semaphores
to SwiftShader when it is built on Linux and Android.

Both are implemented using memfd-backed shared memory regions.

This corresponds to the following extensions:

  VK_KHR_external_memory
  VK_KHR_external_memory_capabilities
  VK_KHR_external_semaphore

DO NOT SUBMIT: This does not report the extensions as supported yet
               and also misses for some of the intermediate structs
	       that they introduced (e.g.
	       VkPhysicalDeviceExternalFormatInfo).

NOTE: This does not try to implement non-external semaphores, i.e.
      those are still a complete no-op. This may be addressed in
      another CL, that could be inserted before this one.

This is required to be able to use SwiftShader has the software Vulkan
backend for the Fuchsia emulator (since the Fuchsia compositor relies
on these extensions).

Another CL will introduce similar support for SwiftShader when it
is built for Fuchsia (which will be used to use SwiftShader as a
Vulkan ICD directly inside a Fuchsia device or VM instead).