SpirvShader: Fix another edge case with phis and loops
If the merge block M contains phis, and M is a loop header of a second loop, we will get assertions that the back edges of the second loop have not yet been generated.

We're only trying to update the phis of M for the lane exits of the first loop, so filter the phi updates to edges from those blocks.

Bug: b/139424793
Change-Id: Idc197fb93c28cdc4b91174a687179f9fc498344a