Add a simple benchmarking tool.
The tool repeatedly runs a the test Vulkan applications listed in the
test list passed as the first argument, varying the number of hardware
threads spawned and used by SwiftShader.

The stdout of the the test applications is parsed for FPS counter messages,
which are averaged for the lifetime of the process, and are reported once all
the tests have completed.

Change-Id: I74dcedfe275ff8cf05e3eae60d82d5c1639a70c0