DO NOT SUBMIT
This is for checking whether dEQP tests 'strictLines' correctly.

Always rasterize lines as a rectangle

This is Vulkan 1.1's "preferred method of line rasterization, and must
be used when the implementation advertises the strictLines limit in
VkPhysicalDeviceLimits as VK_TRUE."

The 'Diamond test convention' code path, which implemented coverage
corresponding to the Bresenham algorithm (except for endpoints, see
b/80135519), is now inactive, but might become of use again for
implementing VK_LINE_RASTERIZATION_MODE_BRESENHAM_EXT so it is kept,
for now.

Bug: b/139755234
Change-Id: I6f943fb7cbae7428740c0ac1a1e6a170ef458ab7