Fix samplerless image fetch
OpImageFetch takes an Image operand, not a SampledImage, and thus we
don't explicitly have a sampler descriptor. However, our
vk::SampledImageDescriptor structure includes both image view descriptor
data and sampler data, and we were relying on the sampler descriptor
data to be present for setting the addressing mode. Other state would
also inadvertently be affected.

This change ensures that the sampler descriptor for fetch operations is
null and we set the addressing mode to repeat but leave other unused
state at their defaults.

Bug: b/139401791
Change-Id: I9f1af35940c2fa9c7d30e771ebdb244f72f5bdbe