Fix mapping NaN inputs for E5B9G9R9 to 0
Section 15.2.1. RGB to Shared Exponent Conversion of the Vulkan 1.1 spec
states "NaN, if supported, is handled as in IEEE 754-2008 minNum() and
maxNum(). That is the result is a NaN is mapped to zero."

This currently does not affect anything because E5B9G9R9 is not a
renderable format. The blitter merely has to support writing this format
for computing the corner texel of cube map borders, which are averaged
from other E5B9G9R9 format texels and thus no NaN can be encountered.

Bug: b/138944025
Change-Id: Ib69aca3af5afee240d9b697c47b2f1a0204fe697