Device: Don't use a global for the cluster count.
Pass the cluster count to the QuadRasterizer instead of using a rather nasty back dependency to the renderer.

This allows the cluster count to be adjusted per-draw without requiring synchronization around the count.

The primary goal of this change is that we can scale the number of clusters based on the complexity of the draw.

Bug: b/139142453
Change-Id: I0379e16568de402f186ee2dd1e8b2346bed30efd