Update patch set 1
Patch Set 1: Code-Review-1

This causes several of these calculations to done for every cluster, instead of most of it getting baked into the routines as constants.

I don't think we should be changing the cluster count on a per-draw case. When we have full command queues we should be able to keep all cores busy. If that's not the case (which I'm very willing to believe) it probably need another solution.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1