Support Windows builds and ANGLE integration.
This CL enables the Windows build integration with ANGLE. Several
instances of "build_with_chromium" in the GN are changed to use
build_overrides instead. Also in several places in the GN support
is added for Windows builds.

Also Add guards for WIN32_LEAN_AND_MEAN, and removes an unused
variable that shows up when building with GN on Windows configs.

Bug: angleproject:3876
Bug: b/140251624
Change-Id: I7a470c7605c2b62420dca44e4797c29983ed7c13