Device/Renderer: Work arounds for MemorySanitizer false-positives
MemorySanitizer has no visibility into writes performed by JIT'd code.

If we have built the project with the MemorySanitizer enabled, memset the Triangle and Primitive structs so the sanitizer has a write to see.

This is only cleared in MemorySanitizer enabled builds, as this adds a non negligable cost to Renderer construction.

Bug: chromium:998457
Change-Id: I3b4f9e48783c6e8a1109a578988567b0a42eb1f9