Check for finite Reactor Float constants
C++ does not have a way to write an infinite or NaN literal,
nor does it allow division by zero as a constant expression.
Thus we should not accept inf or NaN as a Reactor Float constant,
as this would typically idicate a bug, and avoids undefined
behavior.

This also prevents the issue of the LLVM JIT only taking double
values for constructing floating-point constants. During the
conversion from single-precision to double, a signaling NaN can
become a quiet NaN, thus altering its bit pattern. Hence this
assert is also helpful for detecting cases where integers are
being reinterpreted as float and then bitcast to integer again,
which does not guarantee preserving the integer value.

Should inifinty and NaN constants be required, methods like
infinity(), quiet_NaN(), and signaling_NaN() should be added
to the Float class.

Bug: b/140302841
Change-Id: I72cf7e6fe82913cffcb1e0de282bc6839f65226d