[yarn]: Use less atomic variables.
This is an experimental CL to reduce the number of
atomic variables used by various Yarn primitives when
a mutex is already being used to protect their values.

The rationale for this is that on 32-bit ARM, atomic
instructions are very slow (100s or even 200s CPU cycles
each, even with empty caches), while atomic acquire and
release are in the same ballpark (in the non-contended
case). It is thus alleged here that this CL:

- Has the potential to speed up some Yarn synchronization
  operations noticeably for 32-bit ARMv7 machine code.

- Should have negligible effect on x86 and x86_64.

- For 64-bit ARMv8, which includes dedicated hardware
  instructions for load-acquire and store-release,
  effect is unknown at the moment.

DO NOT SUBMIT this CL until its effects have been
properly benchmarked though.

+ Add a small technical note in ConditionVariable::notify_one()
  explaining why both a fiber and a thread will be woken up
  when there are waiters of both types when it is called.

+ Add a note in WaitGroup.h explaining that WaitGroup
  instances can be copied / shared between threads, unlike
  other primitives like ConditionVariable.

Change-Id: Iebdcb38f5dc72981ed237da7d0eebf5fcec83916