Update patch set 4
Patch Set 4:

> > I had a little time to think about this problem in the past few weeks. I think I can implement external semaphores on top of Yarn primitives (and yes, spawning more OS threads will be required, though this will be done in the vk::Semaphore implementation, not Yarn itself). Let me try it this week, and I'll upload some new CLs if this works.
> 
> I've looked into this problem as well, and I have a basic solution available:
> https://swiftshader-review.googlesource.com/c/SwiftShader/+/35573
> I'm open to other suggestions.
> 

That's similar to what I had in mind (except using platform-specific threads to reduce stack size, but that's an optimization detail). Thanks, I'll try to do something on top of this CL then.

> > 
> > Btw, looking at the latest Yarn sources, some questions come to mind:
> > 
> > - Why use a std::shared_ptr<Data> inside the WaitGroup implementation? This seems to be
> >   bad for performance, and I don't see the case for sharing, but maybe I'm missing something.
> 
> It's a very defensive solution to a problem that has bitten us a couple of times.
> See b/133135427 for an example.
> 

Thanks for the pointer. The history is really interesting.

Imho, I find the recommendation on cppreference.com really troublesome. I.e. once you release the mutex and that the predicate is true, there is simply no way to guarantee in the general case that you won't be preempted by another thread that will destroy your condition variable, just before you try to signal it.

Unless you control extremely carefuly the lifecycle of all your threads and synchronization objects, the potential for runtime corruption stille exists. It's not a performance issue, but a correctness one.

The cppreference.com text tries to address that in the last paragraph of the description, but I find the text really unclear. I don't think one should try this optimization ever.

Note that the current implemetation of WaitGroup::done() will signal with the lock acquired anyway, so will always be correct.

It looks like this is not what the std::shared_ptr<Data> is really used for though (I've tried removing it / simplifying the class). Instead, it is used to make the instance copyable, as demonstrated by the example and some of the sources and test files (that's a little bit unusual for a synchronization primitive :-))

> > 
> > - From my past experience, non-relaxed atomic operations are super-slow on ARM (e.g. 100
> >   CPU cycles for each one of them, and degrades even more as the core count increases).
> > 
> >   Since most of the atomics you use are associated with a mutex anyway, it might be
> >   beneficial (at least for this platform) to just use plain integers with appropriate
> >   mutex locks/unlocks (this would save several costly DMB data-cache instructions in the
> >   machine code).
> 
> Aside from the shared_ptr, are there other places you've seen where we're using atomic ops inside mutex locks? I'm sure there's more room for optimization, and contributions are welcome. It's worth mentioning I'm in the process of moving Yarn out to a standalone github project. As part of this work I intend to add some presubmit performance checks to ensure that yarn only gets faster. Integrating CI testing with ARM is always tricky though.
> 

I've uploaded an experimental CL [https://swiftshader-review.googlesource.com/c/SwiftShader/+/35932] to demonstrate what I mean here, hopefully the commit message will explain the rationale for this better than a comment here.

Also another CL to fix the Fuchsia build [https://swiftshader-review.googlesource.com/c/SwiftShader/+/35930], and one to add unit tests to BUILD.gn [https://swiftshader-review.googlesource.com/c/SwiftShader/+/35931]

Let me know when the github repository is up, and I'll be happy to upload pull requests there too :)

> > 
> > - Do you use anything special to benchmark Yarn?
> 
> I've written a tool that repeatedly runs sample apps while modifying the number of threads in use. I've run this tool against the PowerVR sample applications, and charted performance on linux x64.
> Tool: https://swiftshader-review.googlesource.com/c/SwiftShader/+/35288
> Benchmarks: https://docs.google.com/spreadsheets/d/19tiZRL74z2iYeZBFRDedu5gtobFgkzMIQppZbqJiSdg/edit?usp=sharing

Impressive, thanks for these links!

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>