[vulkan]: Implement VkSemaphore using Yarn primitives.
Proper implementation of a VkSemaphore using an std::mutex
and a yarn::ConditionVariable. This should make the implementation
compatible with both fibers and threads at the same time.

A future CL will add platform-specific implementations to cover
external semaphores as well. Which explains why the implementation
details are hidden from VkSemaphore.hpp.

Change-Id: I1db55493a41db0eb60ce9181fe864253db09f4f8