[vulkan] Implement Linux-based external semaphore
DO NOT SUBMIT: Experimental CL to add external semaphore
support for Linux and Android based in a process-shared
pthread mutex + condition variable stored in a shared
memory region backed by memfd_create().

This takes care of waiting for external semaphores
in a background thread when invoked from a Yarn fiber.

DO NOT SUBMIT: This does not expose the implementation
through the Vulkan API though, so this code cannot be
tested right now.

Change-Id: Ifa74c807d3e33914e5a37dd96650c312246c8e4f