[vulkan] Implement Linux-based external semaphore
DO NOT SUBMIT: Experimental CL to add external semaphore
support for Linux and Android based in a process-shared
pthread mutex + condition variable stored in a shared
memory region backed by memfd_create().

This takes care of waiting for external semaphores
in a background thread when invoked from a Yarn fiber.