[fuchsia] Fix Fuchsia build of swiftshader_vulkan_unittests
When building the vulkan library (libvk_swiftshader.so) for Fuchsia,
do not place the stripped binary to $OUT_DIR/swiftshader/ instead of
its default location (i.e. $OUT_DIR/lib/).

That's because the Fuchsia packaging scripts rely on the location
of the unstripped library ($OUT+DIR/lib.unstripped/) to find the
release one (i.e. by replacing 'lib.unstripped' with 'lib').

This resulted in a build error (see bug for details).
This CL works around the issue rather than fixing the scripts.

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1000668

Change-Id: I4137c9edbdc04b3b35a8a43e3f37b9e64fe5b1d7