Log unsupported extensions used in pNext structs
The Vulkan spec states that if an implementation does not support an
extension, then it must ignore all fields other than pNext and pType and
continue on without processing any further information contained in the
extension's structure. We've been using asserts to log these situations
so we can tell which extension structs people are trying to use.

This CL changes our behavior so we silently log these unsupported extensions,
and warn a user once if they're debugging.

Retrying this after updating build files to include added VkStringify.*pp files.

Bug: b/139528538
Change-Id: Icd4ea3f53bf2eec5c2c81079b2ebbaa1b9a855db