Update patch set 5
Patch Set 5: Code-Review-1

(4 comments)

This is nice as an overview of things that need addressing to make the tests pass, but I'd rather not land this as-is. The tests are wrong, and this workaround doesn't seem scalable. We're not under immediate timeline pressure to pass these tests, so please look into doing the proper thing of fixing the tests. Landing this now and trying to revert it later may drag on or leave scars.

Patch-set: 5
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1