Renderer alignment fix
The Renderer needs to be aligned when allocated, otherwise
it leads to all sorts of hard to track issues in the 32 bit
version of the library.

In this instance, Angle's end to end tests on SwiftShader
Vulkan had spurious failures due to having unaligned members
in PixelProcessor, which wasn't aligned because it's used as
a base class of the Renderer class, which wasn't aligned when
allocated. Using an allocator to align the memory fixes the
issue.

I also did a small cleanup of PixelProcessor::setBlendConstant,
as it was a bit hard to debug and changing it this way made it
easier to track what was going on.

Bug b/140888415

Change-Id: Ie5559642456ee4dcb73f9d07c777349f558879ba