Fix intrinsic of getting timer tick for Windows ARM64
src/System/Timer.cpp which was duplicated for Vulkan uses x86 intrinsic __rdtsc
to tick count. This intrinsic is not available on ARM and needs to be fixed.
This issue broke Windows ARM64 build bot and the same issue was fixed for
src/Common/Timer.cpp before (see below links).

https://ci.chromium.org/p/chromium/builders/ci/win32-arm64-rel/7763
https://swiftshader-review.googlesource.com/c/SwiftShader/+/23508

Bug: chromium:893460
Change-Id: I48e3673034650a89fead71447d75f4e486ee68b5