Remove array restriction from EOpEqual, EOpNotEqual
The GLSL-version-specific behavior is already dealt with in earlier
checks; here we can assume we are dealing with GLSL ES 3.00.

The GLSL ES 3.00 spec section 5.7 explicitly calls out == and != as
being supported operations on arrays.

This is an unusual construct in real shaders, but it is used in the
various "complex expression" tests (in an unreachable path).

Test: dEQP-GLES3.functional.shaders.arrays.complex_expression.*
Change-Id: I8a4ece122d604ffb917b94c03c8a086230e9108a