Fix write stencil ops when writeMask != 0xFF
We were checking whether we need to apply the write mask by erroneously
looking at whether the writeMask was non-zero. This meant that when the
write mask was zero, we would compute the stencil value, and store
it, rather than masking it to 0. Instead, what we mean is to apply the
write mask if masking does something: when it's not FF (or 2^n-1, with
n =number of stencil bits).

Bug: b/141484021
Change-Id: I6134289ac8bb13a3502884e4e408f71961814adf