Verify the commit message has a Bug: line
The Git change's commit message should have a line that starts with Bug:
or Issue:, followed by a bug ID. This is important for tracking purposes,
association with follow-up changes, and having a centralized place to
discuss issues.

Note that this change doesn't actuall check that the bug ID is valid.
Also, at this point Kokoro doesn't block submits.

Bug: b/141892461
Change-Id: Idf57534d24842e18dca90ec01e5bb134db612286