Fix how we calculate potential overflows with PBOs
We were previously checking if we could read to the end of the last
row in a pixel buffer object without overrunning the buffer. Instead
we need to check if we could read to the last byte defined by the
parameters and state without overrunning the buffer.

Also added a unittest to ensure that we don't reject valid PBO
operations, and do reject invalid PBO operations.

Bug: 140881221
Change-Id: Id297590b85a86acb88285dc589c41f95a306ae65