Update patch set 18
Patch Set 18: Kokoro-Presubmit+1

Please take a look now, the latest CL allows the implementation of several external memory types to coexist in the same build.

Note that the code is a little bit more complex than necessary because I tried to keep the allocation of the implementation separate from its construction (to respect ComputeRequiredSize). Hence the use of the traits types and the templates.

We can drastically simplify the code if we don't care about that, because we then can have a simple factory function that returns a new allocated-and-constructed instance instead. Let me know if this is desirable.

Patch-set: 18
Reviewer: Gerrit User 5045 <5045@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Kokoro-Presubmit=+1