Update patch set 16
Patch Set 16:

(1 comment)

> Patch Set 16:
> 
> Adding printf()s to the start of each function (signaled is changed after printing): 
> 
>   signal(): signaled: false
>   addRef()
>   tryWait(): signaled: true
>   signal(): signaled: false
>   signal(): signaled: true
>   tryWait(): signaled: true
>   tryWait(): signaled: false
>   wait(): signaled: false
> 
> This seems deterministic. It looks suspicious that we have two signals followed by two tryWaits. If they were interleaved, I'd imagine we wouldn't get stuck.

Interesting, are you sure these correspond to the same semaphore? If so, it explains why the test times out, since the second signal is lost, due to the nature of Vulkan semaphores.

It may be that the test is flaky, e.g. it uses several threads with invalid synchronization, or something like that. That would explain why the test passes on my machine, but not the bot too.

I'm going to have a look at its source code first.

Patch-set: 16
Reviewer: Gerrit User 5045 <5045@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>