Workaround a compiler issue
It looks like clang-cl is clobbering a floating point register
when passing a floating point argument to a DLL, because the
first time glUniform1f() is called, the value is set to 0.0f.
All subsequent glUniform1f() calls work properly.

I changed glUniform1f() calls to glUniform1fv() calls in the
AtanCornerCases test to get the test to pass again, but this
is just masking the compiler issue.

Bug: chromium:1013375
Change-Id: I8363182f26c50cd50c07ccea65ab2eb640a7da76