Fix sRGB conversion not applying generally
We want to allow sRGB conversion always, except when blitting to/from
a buffer. This mimics how it's done in the non-VK path.

I also opportunistically made Options constructors explicit to make it
easier to locate instantiation sites, and renamed convertSRGB to
allowSRGBConversion to make it clear that setting this to true doesn't
mean SRGB conversion will necessarily happen.

Bug: angleproject:4011
Change-Id: I88c229e683d26246b21ef601e066df6eb429e0a6