Update patch set 5
Patch Set 5:

Note that the commit comment is incorrect. I wrote:

"We want to allow sRGB conversion always, except when blitting to/from
a buffer. This mimics how it's done in the non-VK path."

However, in the end, it should read:

"We now also allow sRGB conversion when one of either the source or desination formats is sRGB, regardless of filtering or number of samples."

Patch-set: 5