Avoid sync issues by removing global android swapchain map
Moved some functionality from libVulkan.cpp to VkImage.cpp
in order to be able to remove the global androidSwapchainMap
which may have been causing sync issues.

I tried to make as much code as possible compile on desktop
so that we have a better chance of detecting if a desktop
related change breaks Android specific code.

Bug: b/143902189
Change-Id: Ie9e17a77c2a1705b3194ebd67adeb1a3c0e6213f