[fuchsia] VK_FUCHSIA_buffer_collection extension.
This extension is used to make Fuchsia-specific swapchain
images available to the Vulkan ICD, more precisely:

- The VkBufferCollectionFUCHSIA type points to a
  fuchsia.sysmem.BufferCollection object, which models
  a set of image buffers whose format/size/properties
  is negociated by multiple parties (outside of the
  Vulkan ICD), and allocated through the fuchsia.sysmem
  service.

- The extension allows clients to create a VkDeviceMemory
  instance that is backed by BufferCollection buffer
  (identified by its collection handle + buffer index),
  and to bind it to a VkImage or VkBuffer.

  As usual with Vulkan, this requires the use of
  explicit extension structs when creating the
  memory/image/buffer objects, with matching properties
  that are checked at runtime.

- Note that swapchain management is entirely handled by
  a Vulkan layer on Fuchsia, so there is no need to add
  anything specific to src/WSI/ to support the platform.
  This allows the ICD to work with both a fullscreen
  swapchain or the windowed one provided by the Scenic
  compositor (which one is used is determined by which
  Vulkan layer is loaded when creating the instance, and
  is transparent to the ICD).

This is enough to pass the Fuchsia image-pipe-swapchain-tests
as well as the graphics-compute-vk-unittests (except the
Spinel ones who depend on a whitelist of GPU device IDs to
work properly, which currently does not include SwiftShader).

IMPORTANT: Any rendering test will fail because Fuchsia
prevents mapping pages as executable unless they come from
a kernel-verified signed filesystem page. This is not a
SwiftShader issue, and there are work-arounds that currently
require modifying each individual test and some Fuchsia build
files though.

Bug: b/140419396

Change-Id: If03f1507569b92dad76e5953bd5eb751c34e2648