Output Swiftshader version to logcat
This change allows for outputting the SwiftShader version to Android's
logcat. The version information is as follows:

major.minor.patch.commit_hash

In logcat, when SwiftShader is loaded by ANGLE, this looks like:

11-07 12:17:10.152  9110  9110 I SwiftShader: SwiftShader Version: 5.0.0.9c7bf8093a75
11-07 12:17:10.162  9110  9110 I ANGLE   : Version (2.1.0.a2b1a958e2d8), Renderer (Vulkan 1.1.0(SwiftShader Device (0x0000C0DE)))

This allows developers to verify that they are running the expecting
build of SwiftShader when debugging problems or running conformance
tests.

The version information is logged to logcat when the build flag
"ENABLE_BUILD_VERSION_OUTPUT" is enabled within src/Android.bp.

Bug: b/142828252
Change-Id: Iff773b16a2f3532aa843629ec50519b519bbadac